+++ src/App.js
@@ -8,22 +8,22 @@ import Sponsors from "./pages/Sponsors/Sponsors";
 import Resources from "./pages/Resources";
 
 function App() {
-  const [data, setData] = useState(null);
+  //const [data, setData] = useState(null);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await fetch(`http://localhost:1337/api/resources`);
-        const data = await response.json();
-        setData(data);
-        console.log(data.data[0].attributes.title); //can use this to check if the data is fetched correctly -petrine
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
+  // useEffect(() => {
+  //   const fetchData = async () => {
+  //     try {
+  //       const response = await fetch(`http://localhost:1337/api/resources`);
+  //       const data = await response.json();
+  //       setData(data);
+  //       console.log(data.data[0].attributes.title); //can use this to check if the data is fetched correctly -petrine
+  //     } catch (error) {
+  //       console.error('Error fetching data:', error);
+  //     }
+  //   };
 
-    fetchData();
-  }, []);
+  //   fetchData();
+  // }, []);
 
   return (
     <Router>
@@ -34,7 +34,7 @@ function App() {
         <Routes>
           <Route path="/" exact element={<HomePage />} />
           <Route path="/sponsors" exact element={<Sponsors />} />
-          <Route path="/resources" exact element={<Resources posts={data}/>} />
+          <Route path="/resources" element={<Resources />} />
         </Routes>
       </div>
     </Router>
