+++ strapi/env
@@ -1,0 +1,10 @@
+HOST=0.0.0.0
+PORT=1337
+APP_KEYS=UoBtqAtasLU1TavFKif+Pw==,d7FWrhS7qkezECR52c/DLg==,MC7S44IIKNkvpNCTOBNfpQ==,D0e610KF3IAb1qlPe/mM1g==
+API_TOKEN_SALT=v/YlwNFIK0xBoIzGU32mdg==
+ADMIN_JWT_SECRET=3t+gYjTxAmiBgHmTyMV2Lg==
+TRANSFER_TOKEN_SALT=YhdYSRDRobRPtXlBVQh2vA==
+# Database
+DATABASE_CLIENT=sqlite
+DATABASE_FILENAME=.tmp/data.db
+JWT_SECRET=dRnljOSBUUrzr3z060O2Wg==
